+++ Extra/XCode Templates/VCore (SwiftUI)/ContentView (MVVM).xctemplate/___FILEBASENAME___LocalizedStrings.swift
@@ -1,0 +1,13 @@
+//  ___FILEHEADER___
+
+import Foundation
+import VCore
+
+// MARK: - ___VARIABLE_productName___ Localized Strings
+enum ___VARIABLE_productName___LocalizedStrings: String, LocalizationKeyProvider {
+    // MARK: Cases
+    case title
+
+    // MARK: Localization Key Provider
+    var key: String { "scenes.???.\(rawValue)" }
+}